Add user can delete recording option in tour list

+++ app/controllers/user/recordings_controller.rb
@@ -30,7 +30,7 @@ class User::RecordingsController < User::BaseController
 
   def destroy
     recording = current_user.recordings.find_by(id: params[:id])
-    if recording
+    if recording && recording.tours.empty?
       recording.destroy
     else
       flash[:danger] = "That recording cannot be deleted"
